Remove mostly-duplicate code for Rules which can never trigger
PreviewScrips includes a mostly-duplicated block of code to preview
emails produced by Rules, the mechanism behind Approvals.  However, this
code is unnecessary because /Ticket/Update.html is not accessible for
tickets of type 'approval' (the only objects which could currently
trigger Rules in a stock RT), because /Tickets/autohandler redirects all
but a small number of files under /Tickets/ to /Approvals/ instead.
/Ticket/Update.html is not in the whitelist, and thus cannot render for
Approvals.

Furthermore, the code checks $rule->{hints}{class} to determine if it
should display the rule; no rule ever has a ->{hints} set on it, and
never has, since the introduction of this code in 6d14f535.  As such,
even if /Ticket/Update.html were accessible, or some local customization
made use of Rules for something other than Approvals, they would still
not appear in PreviewScrips.

Remove the duplicated and unused code.