Pass extra context to _GroupingClass
Because we're calling _GroupingClass as a class method (rather than on a
loaded CF object) we limit the information available to _GroupingClass
and anyone overriding it (like RTIR).  In the case of LimitToGrouping
being called on a Ticket during creation, you get two incomplete objects
and really can't make any judgements other than the core "ref $object".

For RTIR, which wants to know what Queue you're creating the ticket in,
passing the ContextObject makes this doable.  If called on a loaded
CustomField object, code should use ACLEquivalenceObjects instead, but
this 'rescues' the case of searching for CFs available to a ticket
during creation by providing more context to users overriding
_GroupingClass.

It turns out to be surprisingly reasonable to do this override in case
you want to segment your groups across tickets (or other objects) in the
system.