INSERT DELAYED provides notable speed benefits on MyISAM
MySQL supports INSERT DELAYED[1] for MyISAM (but not InnoDB) tables.  This
allows the server to defer inserts until it has a good opportunity to
write them, and to write them in bulk.  While there is a small risk of
data loss (if the server is terminated before the data is written) this
poses no problem for an AttachmentsIndex table, for which all inserted
data is trivial to re-generate.

[1] http://dev.mysql.com/doc/refman/5.1/en/insert-delayed.html