Drop sphinx xmlpipe2 output, which was unusable and undocumented
The --xmlpipe2 option to rt-fulltext-indexer was intended to be fed into
the "xmlpipe2" datasource of sphinx.  However, it was never documented.
More importantly, it would not work as intended because the
LastIndexedAttachments attribute that it checks is not set anywhere; as
such, every run would infinitely repeat the same $OPT{limit} elements.

Remove the code.